+++ R/package.R
@@ -66,6 +66,7 @@ ensure_python_initialized <- function(required_module = NULL) {
 
   # remap output streams to R output handlers
   remap_output_streams()
+  set_knitr_python_stdout_hook()
 
   # generate 'R' helper object
   py_inject_r()